imports-main/data_importing/requirements.txt
jupyter
ipykernel
firebase-admin>=6.0.0
pandas>=1.3.0
virtualenv>=20.26.1
pyyaml>=6.0.1
openai>=-1.26.0
blackcellmagic
python-dotenv
ipython
ipywidgets
requests
pyairtable
gspread
oauth2client
unittest2
setuptools
openpyxl
tabula-py
jpype1
fuzzywuzzy
notebook
jupyter_contrib_nbextensions
